requirements.txt
boto3==1.33.3
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-countries==7.5.1
django-extensions==3.2.3
django-filter==23.4
django-mptt==0.15.0
django-on-heroku==1.1.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0